Czyszczenie

+++ src/main/java/prk/ski/jumping/model/domain/TournamentWorldCup.java
@@ -1,7 +1,6 @@
 package prk.ski.jumping.model.domain;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.Objects;
 
 /**
@@ -84,7 +83,7 @@ public class TournamentWorldCup {
     }
 
     public int hashCode() {
-        return (int) Objects.hash(id, date, place, gender, link);
+        return Objects.hash(id, date, place, gender, link);
     }
 
     public String toString() {
